requirements.txt
# sudo apt install libkrb5-dev python-winrm
pywinrm[Kerberos]